Elaborated install info Database if not file

+++ setup.py
@@ -1,20 +1,20 @@
 """A setuptools based setup module.
-
-See:
-https://packaging.python.org/en/latest/distributing.html
 """
 import os
 import urllib.request
 from setuptools import setup, find_packages
 
-rem_file = 'http://download.pisak.org/n_grams.sqlite'
+# ---- Prediction database download ----
 loc_file = os.path.join('res', 'n_grams.sqlite')
 
-def dlProgress(count, blockSize, totalSize):
-      percent = str(int(100*count*blockSize/totalSize))
-      print("\r Downloading " + rem_file + "... {}%".format(percent))
+if not os.path.isfile(loc_file):
+    def dlProgress(count, blockSize, totalSize):
+        percent = str(int(100*count*blockSize / totalSize))
+        print("\r Downloading " + rem_file + "... {}%".format(percent))
 
-urllib.request.urlretrieve(rem_file, loc_file, reporthook=dlProgress)
+    rem_file = 'http://download.pisak.org/n_grams.sqlite'
+    urllib.request.urlretrieve(rem_file, loc_file, reporthook=dlProgress)
+# -------------------------------------
 
 setup(
     name='pisak',
